# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Tutor API",
    "description": "API for AI Tutor, managing students, courses, and AI-generated questions. For more details, visit the [GitHub repository](https://github.com/ShishirShekhar/E-Learning-Platform.git).",
    "version": "1.0.0",
    "contact": {
      "name": "Shishir Shekhar",
      "email": "sspdav02@gmail.com"
    },
    "termOfService": "https://www.google.com",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local server"
    }
  ],
...
  "paths": {
...
    "/api/v1/student/enroll": {
      "post": {
        "tags": [
          "Student"
        ],
        "summary": "Enroll in a course",
        "description": "Endpoint to enroll a student in a course",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Course ID",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "courseId": {
                  "type": "string",
                  "example": "611f0f0e9b3f6d001f2e7d6f"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Enrollment successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Enrollment successful"
                        },
                        "student": {
                          "type": "object",
                          "properties": {
                            "_id": {
                              "type": "string",
                              "example": "60f1b0b3b3b3b30015f1b0b3"
                            },
                            "name": {
                              "type": "string",
                              "example": "Shishir"
                            },
                            "email": {
                              "type": "string",
                              "example": "sspdav02@gmail.com"
                            },
                            "role": {
                              "type": "string",
                              "example": "student"
                            },
                            "courses": {
                              "type": "array",
                              "example": [
                                "611f0f0e9b3f6d001f2e7d6f"
                              ],
                              "items": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    },
                    "error": {}
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Enrollment successful"
                        },
                        "student": {
                          "type": "object",
                          "properties": {
                            "_id": {
                              "type": "string",
                              "example": "60f1b0b3b3b3b30015f1b0b3"
                            },
                            "name": {
                              "type": "string",
                              "example": "Shishir"
                            },
                            "email": {
                              "type": "string",
                              "example": "sspdav02@gmail.com"
                            },
                            "role": {
                              "type": "string",
                              "example": "student"
                            },
                            "courses": {
                              "type": "array",
                              "example": [
                                "611f0f0e9b3f6d001f2e7d6f"
                              ],
                              "items": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    },
                    "error": {}
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Student is already enrolled in this course",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Course ID is required"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Course ID is required"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            },
            "schema": {
              "type": "object",
              "properties": {
                "data": {},
                "error": {
                  "type": "string",
                  "example": "Student is already enrolled in this course"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Student not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Course not found"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Course not found"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            },
            "schema": {
              "type": "object",
              "properties": {
                "data": {},
                "error": {
                  "type": "string",
                  "example": "Student not found"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/student/progress": {
      "get": {
        "tags": [
          "Student"
        ],
        "summary": "View progress",
        "description": "Endpoint to view progress of all courses the student is enrolled in",
        "responses": {
          "200": {
            "description": "Student progress retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "60f1b0b3b3b3b30015f1b0b3"
                        },
                        "name": {
                          "type": "string",
                          "example": "Shishir"
                        },
                        "email": {
                          "type": "string",
                          "example": "sspdav02@gmail.com"
                        },
                        "role": {
                          "type": "string",
                          "example": "student"
                        }
                      }
                    },
                    "error": {}
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "60f1b0b3b3b3b30015f1b0b3"
                        },
                        "name": {
                          "type": "string",
                          "example": "Shishir"
                        },
                        "email": {
                          "type": "string",
                          "example": "sspdav02@gmail.com"
                        },
                        "role": {
                          "type": "string",
                          "example": "student"
                        }
                      }
                    },
                    "error": {}
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Student not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Student not found"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Student not found"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}